Rename to Vector3dLike and implement equals, hashcode, and examinable
